Add all assets related to the logos
These were uploaded in the exact format, hierarchy, extensions, filenames, etc. as provided by the artist. It is customary to keep it that way in cases there needs to be adjustments to be discussed and made with the original artist (except it is usually not in a git repo, but considering the OSS nature of the project, website repo makes a lot of sense for this).